@@ -7,11 +7,11 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "lodash.memoize": "^4.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.6",
-    "react-router-dom": "5.0.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-stripe-checkout": "^2.6.3",
     "redux": "^4.1.2",
     "redux-persist": "^6.0.0",
