@@ -13,14 +13,16 @@ const CheckoutItem = ({quantity, id}) => {
             <div className="image-container">
                 <div className="img"style={{backgroundImage: `url(${imageUrl})`}}></div>
             </div>
-            <span className="name">{name}</span>
-            <span className="quantity">
-                <div className="arrow" onClick={()=>dispatch(removeItemFromCart(id))}>&#10094;</div>
-                <span className="value">{quantity}</span>
-                <div className="arrow" onClick={()=>dispatch(addItemToCart(id))}>&#10095;</div>
-            </span>
-            <span className="price">${price}</span>
-            <div className="remove-button" onClick={()=>dispatch(clearItemFromCart(id))}>&#10005;</div>
+            <div className="properties-container">
+                <span className="name">{name}</span>
+                <span className="quantity">
+                    <div className="arrow" onClick={()=>dispatch(removeItemFromCart(id))}>&#10094;</div>
+                    <span className="value">{quantity}</span>
+                    <div className="arrow" onClick={()=>dispatch(addItemToCart(id))}>&#10095;</div>
+                </span>
+                <span className="price">${price}</span>
+                <div className="remove-button" onClick={()=>dispatch(clearItemFromCart(id))}>&#10005;</div>
+            </div>
         </div>
     )
 }
