@@ -1,6 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
+import { useState } from 'react';
+
 
 /*Style and icons */
 import './header.styles.scss'
@@ -9,6 +11,7 @@ import { ReactComponent as Logo } from '../../logo.svg'
 /*Components */
 import CartIcon from '../cart-icon/cart-icon.component';
 import CartDropdown from '../cart-dropdown/cart-dropdown.component';
+import BurgerIcon from '../burger-icon/burger-icon.component'
 
 /*Selectors */
 import { selectCartHidden } from '../../redux/cart/cart.selectors';
@@ -16,15 +19,29 @@ import { selectCurrentUser } from '../../redux/user/user.selectors';
 
 /*Actions */
 import { signOutStart } from '../../redux/user/user.action'
+import { toggleMenuListShowPropery } from '../../redux/app/app.action'
+
 
 const Header = () => {
     const dispatch = useDispatch();
     const currentUser = useSelector(selectCurrentUser);
     const isCartDropdownHided = useSelector(selectCartHidden);
 
+    const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+
+
+    
+    const screenSizeHandler = () => {
+        setScreenWidth(window.innerWidth);
+    }
+    
+    window.addEventListener("resize", screenSizeHandler); 
+    
     return (
         <div className='header' >
-            <div className='logo-container'><Link to='/' ><Logo/></Link></div>
+            <div className='logo-container' onClick={()=>{dispatch(toggleMenuListShowPropery())}}><Link to='/' ><Logo/></Link></div>
+            {
+            screenWidth >= 500 ?    
             <div className='options'>
                 <Link className='option' to='/shop'>SHOP</Link>
                 <Link className='option' to='/contact'>CONTACT</Link>
@@ -35,7 +52,9 @@ const Header = () => {
                     <CartDropdown/>
                 }
             </div>
-        </div>
+            :
+            <BurgerIcon clickHandler={()=>dispatch(toggleMenuListShowPropery())}/>
+        }</div>
     )
 }
 export default Header;
