@@ -3,17 +3,17 @@ import { compose } from "redux";
 import { connect } from "react-redux";
 import { WithSpinner } from "../with-spinner/with-spinner.component";
 
-import SectionComponent from "./section.component";
+import Section from "./section.component";
 import { selectIsCollectionsInFetching } from "../../redux/shop-data/shop-data.selector";
 
 const mapStateToProps = state => ({
     isCollectionsInFetching: selectIsCollectionsInFetching(state)
 })
 
-const SectionComponentContainer = compose(
+const SectionContainer = compose(
     connect(mapStateToProps),
     WithSpinner
-)(SectionComponent)
+)(Section)
 
-export default SectionComponentContainer
+export default SectionContainer
 
